345 core add support for queryinterface where modulesplugins could facilitate easier communication with each other (#356)
* #345 updated GView to support Query Interface
#346 added initial support for LLMs
* + pin ubuntu 22.04 as build OS
* + pin ubuntu 22.04 as build OS for code-ql

+++ GViewCore/src/View/DissasmViewer/jclass/DissasmJClass.cpp
@@ -28,9 +28,9 @@ bool Instance::DrawDissasmJavaByteCodeZone(DrawLineInfo& dli, DissasmCodeZone* z
 
         auto clazz = creator.create();
         if (clazz) {
-            PopulateZoneTextToDissasmAsmPreCacheLine(zone, "Constant data:", strlen("Constant data:"));
+            PopulateZoneTextToDissasmAsmPreCacheLine(zone, "Constant data:", (uint32)strlen("Constant data:"));
             if (parser.constant_data.empty()) {
-                PopulateZoneTextToDissasmAsmPreCacheLine(zone, "No constant data!", strlen("No constant data!"));
+                PopulateZoneTextToDissasmAsmPreCacheLine(zone, "No constant data!", (uint32)strlen("No constant data!"));
             }
 
             LocalString<64> line;
@@ -40,10 +40,10 @@ bool Instance::DrawDissasmJavaByteCodeZone(DrawLineInfo& dli, DissasmCodeZone* z
             }
         } else {
             const char* msg = "Failed to parse class file, there are still some features in progress!";
-            PopulateZoneTextToDissasmAsmPreCacheLine(zone, msg, strlen(msg));
+            PopulateZoneTextToDissasmAsmPreCacheLine(zone, msg, (uint32)strlen(msg));
         }
 
-        AdjustZoneExtendedSize(zone, zone->asmPreCacheData.cachedAsmLines.size());
+        AdjustZoneExtendedSize(zone, (uint32)zone->asmPreCacheData.cachedAsmLines.size());
 
         zone->isInit = true;
     }
